разбивает проект на модули
