+++ js/modules/util.js
@@ -1,0 +1,21 @@
+const getRandomPositiveFloat = (first, second, digits = 1) => {
+
+  const lower = Math.min(Math.abs(first), Math.abs(second));
+  const upper = Math.max(Math.abs(first), Math.abs(second));
+
+  const result = Math.random() * (upper - lower) + lower;
+
+  return result.toFixed(digits);
+};
+
+const getRandomPositiveInteger = (first, second) => {
+
+  const lower = Math.ceil(Math.min(Math.abs(first), Math.abs(second)));
+  const upper = Math.floor(Math.max(Math.abs(first), Math.abs(second)));
+
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+export {getRandomPositiveFloat, getRandomPositiveInteger};
